requirement.txt
pydobotplus
pyrealsense2
numpy
pupil-apriltags
matplotlib
opencv-python
scipy
PyYAML